backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "insightboard"
version = "1.0.0"
description = "Real-time analytics platform with event-driven architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.23.2",
    "sqlalchemy[asyncio]==2.0.23",
    "alembic==1.12.1",
    "asyncpg==0.28.0",
    "psycopg2-binary==2.9.9",
    "aiosqlite==0.19.0",
    "aiokafka==0.10.0",
    "redis[hiredis]==5.0.1",
    "pydantic[email]==2.4.2",
    "pydantic-settings==2.0.3",
    "python-dotenv==1.0.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "httpx==0.25.1",
    "prometheus-client==0.17.1",
    "typer[all]==0.9.0",
    "rich==13.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-dotenv==0.5.2",
    "black==23.10.1",
    "ruff==0.1.3",
    "mypy==1.6.1",
    "pre-commit==3.5.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
exclude = ["migrations/", "tests/"]
ignore = ["E501", "E402", "N803", "N806", "E722"]

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = "migrations/"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "^app/tests/",
]

[[tool.mypy.overrides]]
module = "app.tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "aiokafka.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --strict-markers --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
fail_under = 83
